requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
qrcode[pil]==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0

# Para desarrollo
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
